requirements.txt
PyYAML==5.4.1
argparse==1.4.0